fix(cli): address review feedback
